pyproject.toml
[project]
name = "xtts-finetune"
version = "0.1.0"
description = "Fine tune xtts via CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.0",
    "anyascii>=0.3.3",
    "bangla>=0.0.5",
    "black>=25.9.0",
    "bnnumerizer>=0.0.2",
    "bnunicodenormalizer>=0.1.7",
    "bokeh>=3.8.0",
    "coqpit==0.0.17",
    "coverage>=7.10.7",
    "cutlet>=0.5.0",
    "cython>=3.1.4",
    "einops>=0.8.1",
    "encodec>=0.1.1",
    "flask>=3.1.2",
    "fsspec>=2025.9.0",
    "g2pkk>=0.1.2",
    "gruut[de,es,fr]>=2.2.3",
    "hangul-romanize>=0.1.0",
    "inflect>=7.5.0",
    "isort>=7.0.0",
    "jamo>=0.4.1",
    "jieba>=0.42.1",
    "librosa>=0.11.0",
    "llvmlite>=0.45.1",
    "matplotlib>=3.10.7",
    "mecab-python3>=1.0.10",
    "mutagen>=1.47.0",
    "nltk>=3.9.2",
    "nose2>=0.15.1",
    "num2words>=0.5.14",
    "numba>=0.62.1",
    "numpy>=1.26.4",
    "openai-whisper>=20250625",
    "packaging>=25.0",
    "pandas<2.0",
    "pylint>=4.0.0",
    "pypinyin>=0.55.0",
    "pysbd>=0.3.4",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "soundfile>=0.13.1",
    "spacy[ja]>=3.8.7",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchvision==0.20.1",
    "tqdm>=4.67.1",
    "trainer==0.0.36",
    "transformers==4.48.2",
    "tts==0.22",
    "umap-learn>=0.5.9.post2",
    "unidecode>=1.4.0",
    "unidic-lite>=1.0.8",
]


[tool.uv]
index-strategy = "unsafe-best-match"

[tool.uv.sources]
torch = { index = "pytorch" }
torchaudio = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
